+++ testapp/src/main/java/testapp/endpoint/ghissues/GH631.java
@@ -1,0 +1,18 @@
+package testapp.endpoint.ghissues;
+
+import act.controller.annotation.UrlContext;
+import org.osgl.http.H;
+import org.osgl.mvc.annotation.GetAction;
+
+/**
+ * Test `@With` on action methods
+ */
+@UrlContext("631")
+public class GH631 extends GithubIssueBase {
+
+    @GetAction("{header}")
+    public String test(String header, H.Request req) {
+        return req.header(header);
+    }
+
+}